# Use 'path' command to check goal in sheep scenario