Builds upon #1536 (merge that one first)

| Before | After |
| --- | --- |
| 3.58s | 3.11s |

The goal-checking had already been pretty well optimized by only performing pathfinding when the player has just placed a fence between two others.  But at least the code is significantly simplified.